requirements.txt
absl-py==1.4.0
aio-pika==8.2.3
aiofiles==23.2.1
aiogram==2.15
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiormq==6.4.2
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
APScheduler==3.9.1.post1
astunparse==1.6.3
async-timeout==4.0.3
attrs==22.1.0
Babel==2.9.1
bidict==0.23.1
blis==0.7.11
boto3==1.34.53
botocore==1.34.53
CacheControl==0.12.14
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.2.1
confection==0.1.4
confluent-kafka==2.3.0
cryptography==41.0.7
cycler==0.12.1
cymem==2.0.8
dask==2022.10.2
dnspython==2.3.0
docopt==0.6.2
email-validator==2.1.1
exceptiongroup==1.2.0
fastapi==0.110.0
fbmessenger==6.0.0
filelock==3.13.1
fire==0.5.0
flatbuffers==23.5.26
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
future==1.0.0
gast==0.4.0
google-auth==2.28.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.62.0
gTTS==2.5.1
h11==0.14.0
h5py==3.10.0
httptools==0.6.1
huggingface-hub==0.21.3
humanfriendly==10.0
idna==3.6
importlib-metadata==7.0.1
jax==0.4.25
jieba==0.42.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.2.0
jsonpickle==3.0.3
jsonschema==4.17.3
keras==2.12.0
kiwisolver==1.4.5
langcodes==3.3.0
libclang==16.0.6
locket==1.0.0
magic-filter==1.0.12
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.5.3
mattermostwrapper==2.2
ml-dtypes==0.3.2
msgpack==1.0.8
multidict==5.2.0
murmurhash==1.0.10
networkx==2.6.3
nltk==3.8.1
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==20.9
pamqp==3.2.1
partd==1.4.1
pathlib-abc==0.1.1
pathy==0.11.0
pillow==10.2.0
playsound==1.3.0
pluggy==1.4.0
portalocker==2.8.2
preshed==3.0.9
prompt-toolkit==3.0.28
protobuf==4.23.3
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.9
pydantic-core==2.16.3
pydot==1.4.2
PyJWT==2.8.0
pykwalify==1.8.0
pymongo==4.3.3
pyparsing==3.1.1
pyreadline3==3.4.1
pyrsistent==0.20.0
python-crfsuite==0.9.10
python-dateutil==2.8.2
python-dotenv==1.0.1
python-engineio==4.9.0
python-multipart==0.0.9
python-socketio==5.11.1
pytz==2022.7.1
pywin32==306
PyYAML==6.0.1
questionary==1.10.0
randomname==0.1.5
rasa==3.6.18
rasa-sdk==3.6.2
redis==4.6.0
regex==2022.10.31
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rocketchat-API==1.30.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.8
s3transfer==0.10.0
sanic==21.12.2
Sanic-Cors==2.0.1
sanic-jwt==1.8.0
sanic-routing==0.7.2
scikit-learn==1.1.3
scipy==1.12.0
sentencepiece==0.1.99
sentry-sdk==1.14.0
simple-websocket==1.0.0
six==1.16.0
sklearn-crfsuite==0.3.6
slack-sdk==3.27.1
smart-open==6.4.0
sniffio==1.3.1
spacy==3.4.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.10.1
SQLAlchemy==1.4.51
srsly==2.4.8
starlette==0.36.3
structlog==23.3.0
structlog-sentry==2.0.3
tabulate==0.9.0
tarsafe==0.0.4
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-hub==0.13.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
terminaltables==3.1.10
thinc==8.1.12
threadpoolctl==3.3.0
tokenizers==0.13.3
toolz==0.12.1
tqdm==4.66.2
transformers==4.26.0
twilio==8.2.2
typer==0.7.0
typing-extensions==4.10.0
typing-utils==0.1.0
tzdata==2024.1
tzlocal==5.2
ujson==5.9.0
urllib3==1.26.18
uvicorn==0.27.1
wasabi==0.10.1
wcwidth==0.2.13
webexteamssdk==1.6.1
websockets==10.4
werkzeug==3.0.1
wrapt==1.14.1
wsproto==1.2.0
yarl==1.9.4
zipp==3.17.0
